Extract CompanyCard component to it's own file
